server/benchmark/requirements.txt
aiohttp>=3.9.0
pyyaml>=6.0
numpy>=1.24.0
psutil>=5.9.0

# Optional: for visualization
matplotlib>=3.7.0
seaborn>=0.12.0

